requirements.txt
Flask
pandas
pdfplumber
tabula-py
fuzzywuzzy
python-Levenshtein
